Demonstrating a rails filter bug
This app demonstrates adding before filters that are standard methods on
the controller versus adding a before filter that references a class
with a `.filter` method.

1. bundle and fire up the application with rails server.
2. visit http://localhost:3000/home and notice the filter (which
references a class) raises, despite the home controller requesting it be
skipped.
3. visit http://localhost:3000/home/new and notice that no exception is
raised. The home controller was able to skip the applicable filter
because it was added with a method name rather than a class.